+++ telemetry/configuration.ts
@@ -71,7 +71,8 @@ export class TelemetryConfiguration implements TelemetryConfig {
     // TelemetryAttribute.HttpServerRequestDuration,
   
     // This not included by default as it has a very high cardinality which could increase costs for users
-    // TelemetryAttribute.FgaClientUser
+    // TelemetryAttribute.FgaClientUser,
+    // TelemetryAttribute.FgaClientRequestBatchCheckSize
   ]);
 
   /**
@@ -97,6 +98,7 @@ export class TelemetryConfiguration implements TelemetryConfig {
     TelemetryAttribute.HttpClientRequestDuration,
     TelemetryAttribute.HttpServerRequestDuration,
     TelemetryAttribute.FgaClientUser,
+    TelemetryAttribute.FgaClientRequestBatchCheckSize,
   ]);
 
   /**